Optimize get candidate job interview frequency
